Update buttons advance part2 n remove part 1
